Packaged up c1c0-movement, added a full setup.py that should be robust enough to not break if/when things change
