+++ AUTHORS
@@ -1,0 +1,3 @@
+CUP Robotics
+ccrt@cornell.edu
+https://cornellcuprobotics.com/
