+++ MANIFEST.in
@@ -1,0 +1,1 @@
+include AUTHORS PACKAGES VERSION c1c0-movement/* locomotion_control/* pid_test/* arms_jetson.py
